style: use css variable for max-width (#154)
* style: use css variable for max-width

Update to use a css variable to determine max width for the
header content. This will allow more flexibility when used
in qiksit.org, platypus and others. Defaults to none.

* fix: set tabindex for sidenav

Set sidenav tab index to -1. this does not interrupt
correct tab order and focus when the mobile menu
is expanded. This will ensure that when the sidenav
is hidden (larger breakpoints) the items in the sidepanl
cannot be tabbed into